Update Ruby dependencies and repos
Since this project had not been touched in [checks watch] seven
years, it was using a very old version of Ruby and old gems.
This was creating a problem with an old version of ActiveRecord
and Postgres meaning the application could no longer read from
the database. Frankly there were probably some other issues too.
Update to a new version of Ruby and get a fresh set of gems in here.

Also swap out Hiera for Puppet Strings in terms of the repos we
look at since Hiera does not see much active development these
days and I like Puppet Strings.