Make description of Puppet vs community graph accurate
Since my GitHub account is no longer associated with Puppet, the
data I have access to has changed, I can now only see if someone
is a memeber of the Puppet GitHub org if they have chosen to
list it publicly meaning that the pie chart of puppet vs community
members is not very accurate if I collect the data as a non-employee.

Update the description of the graph slighty to reflet this.