Add Rake task to import single repo on Sundays only
This is for use in Heroku. This is a bandaid to get around GitHub
API rate limiting. Previously we would import on Sunday for all
repos at once based on a list hard coded in the Rakefile. Now we
need to make too many API requests in one hour, so we'd like to
stagger it one repo at a time. Add this single repo job to allow us
to do this and avoid hitting GitHub's rate limits.

As before, the Sunday only check is needed because Heroku will
only let you schedule daily jobs, but if we really imported data
every day we would have too much uptime and wouldn't remain within
Heroku's free tier.